style(frontend): Refactor code to enforce code style rules
This commit standardizes the code syntax across multiple files to
improve readability and enforce certain code style rules. This
involves changes to import statements, adding semicolons to end of
lines, consistent use of single quotes for string literals, and
reformatting code for better clarity. Additionally, a "
.prettierignore" file has been added to ignore directories and files
that should not be formatted by Prettier.

Refs: #1