refactor(frontend): Remove unused components and hooks, add new ones
This commit removes unused components, test files and hooks
SidebarMenu components and replaces them with new one using shadcn ui.
It includes SidebarMenu and useNavigation. New hooks and components
are added such as useSidebar hook for managing sidebar state and
several new components for the better layout and accordion type
navigation. Because of this commit, the UI of the sidebar changes,
moving from a static menu to a dynamic one.

Refs: #1